resolves #3
Censor obtained was tested in different conditions of soil of the plant. The test included using the censor in almost every part of the soil both when the soil was newly watered and when it was dry. The results were within expectations. With these test we have completed the task for choosing the best soil- humidity censor and the implementation of it for micro-controller.